-add better command validation -add account deletion -add tracker commands -add auto login system -change accounts json structure -add authUser -add tracker to user_data in json -add logout check -sync json on logout -deextract help functions -update help with more commands -change some help messages

+++ src/main.py
@@ -1,29 +1,26 @@
 import command.help
 import command.auth
+import command.tracker
+from module.json_handler import current_user_json_handler
+
+COMMAND_CONSTRAINTS = {
+    "help": {"arg_count": [0, 1]},
+    "login": {"arg_count": [2]},
+    "register": {"arg_count": [2]},
+    "logout": {"arg_count": [0]},
+    "delete": {"arg_count": [0]},
+    "tracker_add_task": {"arg_count": [1]},
+    "tracker_remove_task": {"arg_count": [1]},
+    "tracker_reset_task_time": {"arg_count": [1]},
+    "tracker_rename_task": {"arg_count": [2]},
+}
 
 def validateCommand(*args):
-    match args[0]:
-        case "help":
-            if len(args) == 1 or len(args) == 2:
-                return True
+    command = args[0]
+    if command not in COMMAND_CONSTRAINTS:
+        return False
 
-        case "login":
-            if len(args) == 3:
-                return True
-
-        case "register":
-            if len(args) == 3:
-                return True
-
-        case "logout":
-            if len(args) == 1:
-                return True
-
-        case "delete":
-            if len(args) == 1:
-                return True
-
-    return False
+    return len(args) - 1 in COMMAND_CONSTRAINTS[command]["arg_count"]
 
 def exeCommand(*args):
     if not validateCommand(*args):
@@ -44,11 +41,36 @@ def exeCommand(*args):
             command.auth.logout()
 
         case "delete":
-            command.auth.delete()
+            password = input("Enter you password to proceed: ")
+            if not command.auth.authUser(current_user_json_handler.data["username"], password):
+                print("Invalid password!")
+                return
+
+            confirm = input("Are you sure you want to delete your account? (y/n): ")
+            if confirm.lower() == 'y':
+                command.auth.delete()
+
+        case "tracker_add_task":
+            command.tracker.addTask(*args[1:])
+
+        case "tracker_remove_task":
+            command.tracker.removeTask(*args[1:])
+
+        case "tracker_reset_task_time":
+            command.tracker.resetTaskTime(*args[1:])
+
+        case "tracker_rename_task":
+            command.tracker.renameTask(*args[1:])
+
+        case _:
+            raise Exception("Unknown Error! command unmatched")
 
 def main():
     print("Productivity Tracker App(no gui)\n")
     print("'help' to list all commands")
+
+    if current_user_json_handler.data['username'] is not None:
+        print(f"Logged in as {current_user_json_handler.data['username']}\n")
 
     while True:
         command = input("command: ")
